Split action mapper builder, update tests and prevent overrides from changing rule tags
