+++ tests/unit/configuration/configuration_manager_test.cpp
@@ -1,0 +1,19 @@
+// Unless explicitly stated otherwise all files in this repository are
+// dual-licensed under the Apache-2.0 License or BSD-3-Clause License.
+//
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2021 Datadog, Inc.
+
+#include "common/gtest_utils.hpp"
+#include "configuration/actions_parser.hpp"
+#include "configuration/common/common.hpp"
+#include "configuration/configuration_manager.hpp"
+#include "parameter.hpp"
+
+using namespace ddwaf;
+
+namespace {
+
+TEST(TestConfigurationManager, EmptyConfig) {}
+
+} // namespace
